+++ examples/pong.rs
@@ -90,14 +90,14 @@ fn main() {
     pause_message.set_string("Welcome to SFML pong!\nPress space to start the game");
 
     // Define the paddles properties
-    let mut ai_timer = Clock::start();
+    let mut ai_timer = Clock::start().unwrap();
     let ai_time = Time::seconds(0.0333);
     let paddle_speed = 400.;
     let mut right_paddle_speed = 0.;
     let mut ball_speed = 400.;
     let mut ball_angle = 0.;
 
-    let mut clock = Clock::start();
+    let mut clock = Clock::start().unwrap();
     let mut is_playing = false;
     let mut up = false;
     let mut down = false;