+++ examples/shader.rs
@@ -46,7 +46,8 @@ impl<'t> Drawable for Pixelate<'t> {
 impl<'t> Effect for Pixelate<'t> {
     fn update(&mut self, _t: f32, x: f32, y: f32) {
         self.shader
-            .set_uniform_float("pixel_threshold", (x + y) / 30.0);
+            .set_uniform_float("pixel_threshold", (x + y) / 30.0)
+            .unwrap();
     }
     fn name(&self) -> &str {
         "pixelate"
@@ -110,11 +111,13 @@ impl<'fo> Drawable for WaveBlur<'fo> {
 
 impl<'fo> Effect for WaveBlur<'fo> {
     fn update(&mut self, t: f32, x: f32, y: f32) {
-        self.shader.set_uniform_float("wave_phase", t);
+        self.shader.set_uniform_float("wave_phase", t).unwrap();
         self.shader
-            .set_uniform_vec2("wave_amplitude", Vector2f::new(x * 40., y * 40.));
+            .set_uniform_vec2("wave_amplitude", Vector2f::new(x * 40., y * 40.))
+            .unwrap();
         self.shader
-            .set_uniform_float("blur_radius", (x + y) * 0.008);
+            .set_uniform_float("blur_radius", (x + y) * 0.008)
+            .unwrap();
     }
     fn name(&self) -> &str {
         "wave + blur"
@@ -168,12 +171,17 @@ impl Effect for StormBlink {
     fn update(&mut self, t: f32, x: f32, y: f32) {
         let radius = 200. + t.cos() * 150.;
         self.shader
-            .set_uniform_vec2("storm_position", Vector2f::new(x * 800., y * 600.));
+            .set_uniform_vec2("storm_position", Vector2f::new(x * 800., y * 600.))
+            .unwrap();
         self.shader
-            .set_uniform_float("storm_inner_radius", radius / 3.);
-        self.shader.set_uniform_float("storm_total_radius", radius);
+            .set_uniform_float("storm_inner_radius", radius / 3.)
+            .unwrap();
         self.shader
-            .set_uniform_float("blink_alpha", 0.5 + (t * 3.).cos() * 0.25);
+            .set_uniform_float("storm_total_radius", radius)
+            .unwrap();
+        self.shader
+            .set_uniform_float("blink_alpha", 0.5 + (t * 3.).cos() * 0.25)
+            .unwrap();
     }
     fn name(&self) -> &str {
         "storm + blink"
@@ -207,7 +215,7 @@ impl<'t> Edge<'t> {
 
         let mut shader =
             Shader::from_file(example_res!("edge.frag"), ShaderType::Fragment).unwrap();
-        shader.set_uniform_current_texture("texture");
+        shader.set_uniform_current_texture("texture").unwrap();
 
         Self {
             surface,
@@ -233,7 +241,8 @@ impl<'t> Drawable for Edge<'t> {
 impl<'t> Effect for Edge<'t> {
     fn update(&mut self, t: f32, x: f32, y: f32) {
         self.shader
-            .set_uniform_float("edge_threshold", 1. - (x + y) / 2.);
+            .set_uniform_float("edge_threshold", 1. - (x + y) / 2.)
+            .unwrap();
         let entities_len = self.entities.len() as f32;
 
         for (i, en) in self.entities.iter_mut().enumerate() {
@@ -291,7 +300,7 @@ fn main() {
     let mut instructions = Text::new(msg, &font, 20);
     instructions.set_position((280., 555.));
     instructions.set_fill_color(Color::rgb(80, 80, 80));
-    let clock = Clock::start();
+    let clock = Clock::start().unwrap();
 
     while window.is_open() {
         while let Some(event) = window.poll_event() {