+++ src/audio/capture.rs
@@ -213,7 +213,7 @@ impl<'a, R: SoundRecorder> SoundRecorderDriver<'a, R> {
     /// It can be called on the fly (i.e: while recording).
     /// If you do so while recording and opening the device fails, it stops the recording.
     pub fn set_device(&mut self, name: &str) -> SfResult<()> {
-        let name = CString::new(name).unwrap();
+        let name = CString::new(name).into_sf_result()?;
         let success = unsafe { sfSoundRecorder_setDevice(self.ffi_handle.as_ptr(), name.as_ptr()) };
         success.into_sf_result()
     }
@@ -310,7 +310,7 @@ impl SoundBufferRecorder {
     /// It can be called on the fly (i.e: while recording).
     /// If you do so while recording and opening the device fails, it stops the recording.
     pub fn set_device(&mut self, name: &str) -> SfResult<()> {
-        let name = CString::new(name).unwrap();
+        let name = CString::new(name).into_sf_result()?;
         let success =
             unsafe { sfSoundBufferRecorder_setDevice(self.ffi_handle.as_ptr(), name.as_ptr()) };
         success.into_sf_result()