+++ src/graphics/texture.rs
@@ -172,7 +172,7 @@ impl Texture {
     /// # Arguments
     /// * filename - Path of the image file to load
     pub fn load_from_file(&mut self, filename: &str, area: IntRect) -> SfResult<()> {
-        let c_str = CString::new(filename).unwrap();
+        let c_str = CString::new(filename).into_sf_result()?;
         unsafe { ffi::sfTexture_loadFromFile(self, c_str.as_ptr(), area).into_sf_result() }
     }