+++ src/system/clock.rs
@@ -1,4 +1,6 @@
-use crate::{ffi::system as ffi, sf_box::Dispose, system::Time, SfBox};
+use crate::{
+    ffi::system as ffi, sf_box::Dispose, system::Time, IntoSfResult as _, SfBox, SfResult,
+};
 pub use ffi::sfClock as Clock;
 
 impl Dispose for Clock {
@@ -11,9 +13,8 @@ impl Dispose for Clock {
 
 impl Clock {
     /// Creates a new Clock and starts it automatically.
-    #[must_use]
-    pub fn start() -> SfBox<Self> {
-        unsafe { SfBox::new(ffi::sfClock_new()).unwrap() }
+    pub fn start() -> SfResult<SfBox<Self>> {
+        unsafe { SfBox::new(ffi::sfClock_new()) }.into_sf_result()
     }
 
     /// Gets the elapsed time.
@@ -33,12 +34,5 @@ impl Clock {
     /// It also returns the time elapsed since the clock was started.
     pub fn restart(&mut self) -> Time {
         unsafe { Time::from_raw(ffi::sfClock_restart(self)) }
-    }
-}
-
-impl Default for SfBox<Clock> {
-    /// Equivalent to `Clock::start()`.
-    fn default() -> Self {
-        Clock::start()
     }
 }